:fire:  Rename HomePageViews to HomePageView
