+++ src/views/HomePageView/StatsSection/StatsSection.styles.js
@@ -10,25 +10,28 @@ import { fontWeights } from '../../../styles/typography';
 import breakpoints from '../../../styles/breakpoints';
 import breakpointsRange from '../../../utils/breakpointsRange';
 
+export const SectionContainer = styled.section`
+  ${breakpointsRange(
+    [
+      { prop: 'paddingTop', sizes: [80, 173], bases: [16, 20] },
+      { prop: 'paddingBottom', sizes: [80, 130], bases: [16, 20] },
+    ],
+    breakpoints.spacings
+  )};
+`;
+
 export const StatsIntro = styled.p`
-  margin-top: 130px;
-  margin-bottom: 130px;
+  margin-top: 0;
 
   color: ${colors.bleu80};
   font-weight: ${fontWeights.bold};
 
   ${breakpointsRange(
-    [{ prop: 'fontSize', sizes: [24, 40], bases: [16, 20] }],
-    breakpoints.spacings
-  )};
-
-  ${breakpointsRange(
-    [{ prop: 'lineHeight', sizes: [29, 48], bases: [24, 40], unit: '' }],
-    breakpoints.spacings
-  )};
-
-  ${breakpointsRange(
-    [{ prop: 'marginBottom', sizes: [80, 130], bases: [16, 20] }],
+    [
+      { prop: 'fontSize', sizes: [24, 40], bases: [16, 20] },
+      { prop: 'lineHeight', sizes: [29, 48], bases: [24, 40], unit: '' },
+      { prop: 'marginBottom', sizes: [80, 130], bases: [16, 20] },
+    ],
     breakpoints.spacings
   )};
