+++ src/views/HomePageView/StatsSection/index.js
